rename `fixer` module to `enumerate`
