+++ src/main/java/edu/wisc/regfixer/enumerate/Grafter.java
@@ -1,4 +1,4 @@
-package edu.wisc.regfixer.fixer;
+package edu.wisc.regfixer.enumerate;
 
 import java.util.List;
 import java.util.LinkedList;
