+++ src/main/java/edu/wisc/regfixer/enumerate/Job.java
@@ -1,4 +1,4 @@
-package edu.wisc.regfixer.fixer;
+package edu.wisc.regfixer.enumerate;
 
 import java.util.List;
 import edu.wisc.regfixer.parser.RegexNode;
