+++ src/main/java/edu/wisc/regfixer/enumerate/Partials.java
@@ -1,4 +1,4 @@
-package edu.wisc.regfixer.fixer;
+package edu.wisc.regfixer.enumerate;
 
 import java.util.Arrays;
 import java.util.LinkedList;
