+++ src/main/java/edu/wisc/regfixer/util/ReportStream.java
@@ -4,7 +4,7 @@ import java.io.File;
 import java.io.OutputStream;
 import java.io.PrintStream;
 
-import edu.wisc.regfixer.fixer.Range;
+import edu.wisc.regfixer.enumerate.Range;
 import edu.wisc.regfixer.parser.RegexNode;
 
 public class ReportStream extends PrintStream {
